fix(labo): cast labo_id to integer in labo_ingredient_selections insert (42804)
PostgreSQL cannot infer $1 type in INSERT...SELECT context, causing type mismatch
error when labo_id column expects integer. Add explicit ::integer cast.

+++ src/controllers/laboController.js
@@ -90,7 +90,7 @@ const createLabo = async (req, res) => {
         // Auto-import ingredients from assigned activities into the labo
         await pool.query(
           `INSERT INTO labo_ingredient_selections (labo_id, ingredient_id)
-           SELECT DISTINCT $1, ais.ingredient_id
+           SELECT DISTINCT $1::integer, ais.ingredient_id
            FROM activite_ingredient_selections ais
            WHERE ais.activite_id = ANY($2::int[])
            ON CONFLICT DO NOTHING`,
